create versionless fixtures file
